+++ src/FPath/Rel.hs
@@ -22,12 +22,16 @@ import Data.Either    ( Either( Right ) )
 import Data.Eq        ( Eq )
 import Data.Function  ( ($), id )
 import Data.Maybe     ( Maybe( Just, Nothing ) )
+import Data.String    ( String )
 import Data.Typeable  ( Proxy( Proxy ), TypeRep, typeRep )
+import System.Exit    ( ExitCode )
+import System.IO      ( IO )
 import Text.Show      ( Show )
 
 -- data-textual ------------------------
 
-import Data.Textual  ( Printable, toText )
+import Data.Textual  ( Printable( print ), Textual( textual )
+                     , fromString, toString, toText )
 
 -- lens --------------------------------
 
@@ -35,13 +39,19 @@ import Control.Lens.Prism  ( Prism', prism' )
 
 -- more-unicode ------------------------
 
-import Data.MoreUnicode.Functor  ( (⊳) )
-import Data.MoreUnicode.Lens     ( (##) )
-import Data.MoreUnicode.Tasty    ( (≟) )
+import Data.MoreUnicode.Applicative  ( (∤) )
+import Data.MoreUnicode.Functor      ( (⊳) )
+import Data.MoreUnicode.Lens         ( (⫥), (⩼) )
+import Data.MoreUnicode.Natural      ( ℕ )
+import Data.MoreUnicode.Tasty        ( (≟) )
 
-  -- mtl ---------------------------------
+-- mtl ---------------------------------
 
 import Control.Monad.Except  ( MonadError )
+
+-- parsers -----------------------------
+
+import Text.Parser.Combinators  ( try )
 
 -- tasty -------------------------------
 
@@ -51,6 +61,10 @@ import Test.Tasty  ( TestTree, testGroup )
 
 import Test.Tasty.HUnit  ( testCase )
 
+-- tasty-plus --------------------------
+
+import TastyPlus  ( runTestsP, runTestsReplay, runTestTree )
+
 -- text --------------------------------
 
 import Data.Text  ( last, null )
@@ -59,11 +73,13 @@ import Data.Text  ( last, null )
 --                     local imports                      --
 ------------------------------------------------------------
 
-import FPath.RelDir            ( AsRelDir( _RelDir ), RelDir, reldir )
+import FPath.AsFilePath        ( AsFilePath( filepath ) )
 import FPath.Error.FPathError  ( AsFPathError, FPathError, __FPathEmptyE__ )
 import FPath.Parseable         ( Parseable( parse ) )
+import FPath.RelDir            ( AsRelDir( _RelDir ), RelDir, reldir )
 import FPath.RelFile           ( AsRelFile( _RelFile ), RelFile
                                , relfile )
+import FPath.T.FPath.TestData  ( r0, r1, r2, r3, rf1, rf2, rf3, rf4 )
 
 --------------------------------------------------------------------------------
 
@@ -88,6 +104,55 @@ instance AsRelFile Rel where
 
 ----------------------------------------
 
+instance Printable Rel where
+  print (RelD f) = print f
+  print (RelF f) = print f
+
+instance Textual Rel where
+  textual = try (RelD ⊳ textual) ∤ RelF ⊳ textual
+
+----------------------------------------
+
+instance AsFilePath Rel where
+  filepath = prism' toString fromString
+
+filepathTests ∷ TestTree
+filepathTests =
+  let nothin' = Nothing ∷ Maybe Rel
+      fail s  = testCase s $ nothin' ≟ s ⩼ filepath
+   in testGroup "filepath"
+            [ testCase "r0" $ "./"     ≟ RelD r0 ⫥ filepath
+            , testCase "r1" $ "r/"     ≟ RelD r1 ⫥ filepath
+            , testCase "r2" $ "r/p/"   ≟ RelD r2 ⫥ filepath
+            , testCase "r3" $ "p/q/r/" ≟ RelD r3 ⫥ filepath
+
+            , testCase "r0" $ Just (RelD r0) ≟ "./"     ⩼ filepath
+            , testCase "r1" $ Just (RelD r1) ≟ "r/"     ⩼ filepath
+            , testCase "r2" $ Just (RelD r2) ≟ "r/p/"   ⩼ filepath
+            , testCase "r3" $ Just (RelD r3) ≟ "p/q/r/" ⩼ filepath
+
+            , testCase "rf1" $ "r.e"       ≟ RelF rf1 ⫥ filepath
+            , testCase "rf2" $ "r/p.x"     ≟ RelF rf2 ⫥ filepath
+            , testCase "rf3" $ "p/q/r.mp3" ≟ RelF rf3 ⫥ filepath
+            , testCase "rf4" $ ".x"        ≟ RelF rf4 ⫥ filepath
+
+            , testCase "rf1" $ Just (RelF rf1) ≟ "r.e"       ⩼ filepath
+            , testCase "rf2" $ Just (RelF rf2) ≟ "r/p.x"     ⩼ filepath
+            , testCase "rf3" $ Just (RelF rf3) ≟ "p/q/r.mp3" ⩼ filepath
+            , testCase "rf4" $ Just (RelF rf4) ≟ ".x"        ⩼ filepath
+
+            , fail "/etc"
+            , fail "/etc/pam.d/"
+            , fail "/etc//pam.d/"
+            , fail "\0etc"
+            , fail "etc\0"
+            , fail "e\0c"
+
+            , fail "/etc/pam.d"
+            ]
+
+----------------------------------------
+
 relpathT ∷ TypeRep
 relpathT = typeRep (Proxy ∷ Proxy Rel)
 
@@ -100,13 +165,9 @@ instance Parseable Rel where
                 '/' → RelD ⊳ parse  t
                 _   → RelF ⊳ parse t
 
---------------------------------------------------------------------------------
---                                   tests                                    --
---------------------------------------------------------------------------------
-
 parseRelTests ∷ TestTree
 parseRelTests =
-  let success d f t = testCase t $ Right (d ## f) ≟ parse @Rel @FPathError t
+  let success d f t = testCase t $ Right (d ⫥ f) ≟ parse @Rel @FPathError t
    in testGroup "parseRel"
                 [ success [reldir|./|]         _RelDir "./"
                 , success [reldir|etc/|]       _RelDir "etc/"
@@ -114,7 +175,25 @@ parseRelTests =
                 ]
 
 
+--------------------------------------------------------------------------------
+--                                   tests                                    --
+--------------------------------------------------------------------------------
+
 tests ∷ TestTree
-tests = testGroup "Rel" [ parseRelTests ]
+tests = testGroup "Rel" [ filepathTests, parseRelTests ]
                 
+
+----------------------------------------
+
+_test ∷ IO ExitCode
+_test = runTestTree tests
+
+--------------------
+
+_tests ∷ String → IO ExitCode
+_tests = runTestsP tests
+
+_testr ∷ String → ℕ → IO ExitCode
+_testr = runTestsReplay tests
+
 -- that's all, folks! ----------------------------------------------------------