+++ changelog.md
@@ -1,3 +1,11 @@
+1.0.1.0 2019-10-05
+==================
+- add Printable,Textual,AsFilePath instances for File,Dir,Abs,Rel,FPath
+
+1.0.0.1 2019-09-25
+==================
+- change cabal requirements to use >= to allow for build updates
+
 1.0.0.0 2019-09-22
 ==================
 - first release